requirements.txt
pathvalidate>=3.2.0
PyYAML>=6.0.1
